Start working on the whisper thread - lots more to do
