+++ crates/speech/src/speech.rs
@@ -6,10 +6,6 @@ use std::sync::Arc;
 
 mod whisper_thread;
 
-const WHISPER_MODEL_NAME: &str = "ggml-base.en.bin";
-const TARGET_SAMPLE_RATE: usize = 16_000;
-const HIGH_PASS_CUTOFF_HZ: f32 = 80.0;
-
 actions!(
     speech,
     [
@@ -34,20 +30,6 @@ pub enum TranscriberThreadState {
     Idle,
     Listening,
     Transcribing,
-}
-
-pub struct TranscriptionStream {
-    receiver: Receiver<String>,
-}
-
-impl TranscriptionStream {
-    fn new(receiver: Receiver<String>) -> Self {
-        Self { receiver }
-    }
-
-    pub async fn recv(&mut self) -> Option<String> {
-        self.receiver.recv().await.ok()
-    }
 }
 
 pub struct NotificationStream {
@@ -128,13 +110,25 @@ impl Speech {
         sinks.retain(|subscriber| subscriber.try_send(value.clone()).is_ok());
     }
 
-    fn toggle_listening(&mut self) {
+    fn toggle_listening(&mut self, cx: &mut App) {
         let mut state = self.state.lock();
-        if *state == TranscriberThreadState::Listening {
-            *state = TranscriberThreadState::Idle;
+        if let Some(thread_handle) = self.task.take() {
+            *state = TranscriberThreadState::Disabled;
+            drop(state);
+            thread_handle.join().unwrap_or_else(|_| warn!("Failed to join speech thread"));
             info!("Speech listening stopped");
         } else {
             *state = TranscriberThreadState::Listening;
+            drop(state);
+
+            let transcription_sender = self.transcription_sender.clone();
+            let notification_sender = self.notification_sender.clone();
+            self.task = Some(Speech::run_transcription_loop(
+                self.state.clone(),
+                transcription_sender,
+                notification_sender,
+                cx,
+            ));
             info!("Speech listening started");
         }
     }
@@ -147,92 +141,24 @@ impl Speech {
     ) -> std::thread::JoinHandle<()> {
         info!("Launching transcription loop");
         std::thread::spawn(move || {
-            if let Err(err) =
-                whisper_thread::transcription_loop_body(state, transcription_sender, notification_sender)
-            {
+            if let Err(err) = whisper_thread::transcription_loop_body(
+                state,
+                transcription_sender,
+                notification_sender,
+            ) {
                 error!("error in transcription loop: {}", err);
             }
         })
     }
-
-}
-
-fn downmix_multi_channel(data: &[f32], channels: usize) -> Vec<f32> {
-    if channels <= 1 {
-        return data.to_vec();
-    }
-    let mut mono = Vec::with_capacity(data.len() / channels + 1);
-    for frame in data.chunks(channels) {
-        let sum: f32 = frame.iter().copied().sum();
-        mono.push(sum / channels as f32);
-    }
-    mono
-}
-
-fn resample_to_target(chunk: &[f32], ratio: f32) -> Vec<f32> {
-    if (ratio - 1.0).abs() < f32::EPSILON || chunk.is_empty() {
-        return chunk.to_vec();
-    }
-    let output_len = ((chunk.len() as f32) * ratio).ceil() as usize;
-    let mut output = Vec::with_capacity(output_len.max(1));
-    for i in 0..output_len {
-        let src_pos = i as f32 / ratio;
-        let idx = src_pos.floor() as usize;
-        let frac = src_pos - idx as f32;
-        let next_idx = if idx + 1 < chunk.len() { idx + 1 } else { idx };
-        let a = chunk[idx];
-        let b = chunk[next_idx];
-        output.push(a * (1.0 - frac) + b * frac);
-    }
-    output
-}
-
-fn apply_high_pass_filter(
-    samples: &mut [f32],
-    prev_input: &mut f32,
-    prev_output: &mut f32,
-    cutoff_hz: f32,
-) {
-    let rc = 1.0 / (2.0 * std::f32::consts::PI * cutoff_hz);
-    let dt = 1.0 / TARGET_SAMPLE_RATE as f32;
-    let alpha = rc / (rc + dt);
-    for sample in samples.iter_mut() {
-        let output = alpha * (*prev_output + *sample - *prev_input);
-        *prev_input = *sample;
-        *prev_output = output;
-        *sample = output;
-    }
 }
 
 pub fn init(cx: &mut App) {
-    let mut speech = Speech::new(cx);
-    let state = speech.state.clone();
-    let transcription_sender = speech.transcription_sender.clone();
-    let notification_sender = speech.notification_sender.clone();
-    speech.task = Some(Speech::run_transcription_loop(
-        state,
-        transcription_sender,
-        notification_sender,
-        cx,
-    ));
+    let speech = Speech::new(cx);
     cx.set_global(speech);
 
     cx.on_action(|_: &ToggleDictationChannel, cx| {
-        Speech::update_global(cx, |speech, _| {
-            speech.toggle_listening();
+        Speech::update_global(cx, |speech, cx| {
+            speech.toggle_listening(cx);
         });
     });
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_new() {
-        // This test needs a proper app context to run now.
-        // Disabling for now, will be replaced by a functional test later.
-        // let speech = Speech::new();
-        // assert!(matches!(speech.state, SpeechState::Idle));
-    }
 }
